+++ src/types/staff.d.ts
@@ -1,0 +1,8 @@
+export type Staff = {
+    id: number;
+    firstName: string;
+    lastName: string;
+    firstNameKana: string;
+    lastNameKana: string;
+    email: string;
+};
